String measurements now include last line
Before, lines were only identified by the use of \n. This meant that the
last line was excluded from all measurements. By including the null
character, \0, in what signifies a line, we now include the last line.
That's because the null character is used to terminate strings, and so
it can be found at the end of the last line - \n or not.